# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.FileDeleteRequest:
    properties:
      key:
        description: S3对象键
        example: uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
    required:
    - key
    type: object
  dto.FileDeleteResponse:
    properties:
      key:
        description: 已删除的S3对象键
        example: uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
      message:
        description: 删除结果消息
        example: File deleted successfully
        type: string
    type: object
  dto.FileInfoResponse:
    properties:
      filename:
        description: 原始文件名
        example: avatar.jpg
        type: string
      key:
        description: S3对象键
        example: uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
      mime_type:
        description: MIME类型
        example: image/jpeg
        type: string
      size:
        description: 文件大小（字节）
        example: 1024000
        type: integer
      url:
        description: 文件访问URL
        example: https://your-bucket.s3.us-east-1.amazonaws.com/uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
    type: object
  dto.FileUploadResponse:
    properties:
      filename:
        description: 原始文件名
        example: avatar.jpg
        type: string
      key:
        description: S3对象键
        example: uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
      mime_type:
        description: MIME类型
        example: image/jpeg
        type: string
      size:
        description: 文件大小（字节）
        example: 1024000
        type: integer
      uploaded_at:
        description: 上传时间
        type: string
      url:
        description: 文件访问URL
        example: https://your-bucket.s3.us-east-1.amazonaws.com/uploads/2024/01/15/550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
host: api-dev.718ai.cn
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: AI API Gateway provides unified access to multiple AI services including
    OpenAI-compatible chat completions, Midjourney image generation, and more.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: AI API Gateway
  version: 1.0.0
paths:
  /api/files/{key}:
    get:
      description: 根据文件键获取文件信息和访问URL
      parameters:
      - description: 文件键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileInfoResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 获取文件信息
      tags:
      - 文件管理
  /api/files/delete:
    delete:
      consumes:
      - application/json
      description: 根据文件键删除S3中的文件
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FileDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDeleteResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 删除S3文件
      tags:
      - 文件管理
  /api/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到S3存储，支持图片、PDF等多种文件类型
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileUploadResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/dto.Response'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/dto.Response'
        "415":
          description: 不支持的文件类型
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 上传文件到S3
      tags:
      - 文件管理
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token for API authentication. Format: ''Bearer {token}'''
    in: header
    name: Authorization
    type: apiKey
  MJApiSecret:
    description: API secret for Midjourney endpoints
    in: header
    name: mj-api-secret
    type: apiKey
swagger: "2.0"